+++ packages/utils/test/withIdRef.test.ts
@@ -34,4 +34,8 @@ describe('withIdRefPrefix()', () => {
 
     expect(withIdRefPrefix(schema)).toEqual(schema);
   });
+  it('should handle null schema', () => {
+    const schema: RJSFSchema = null;
+    expect(withIdRefPrefix(schema)).toEqual(schema);
+  });
 });